pyproject.toml
[project]
name = "configator"
version = "3000.0.0"
description = "Load config values from 1Password with this module."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
  "onepassword-sdk==0.3.*",
  "pydantic==2.*",
  "sorcery==0.2.*",
  "structlog==25.*",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.coverage.run]
relative_files = true

[tool.mypy]
python_version = "3.12"
disallow_any_generics = true
disallow_untyped_defs = true
enable_error_code = "ignore-without-code,redundant-expr,truthy-bool,unused-awaitable"
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
  "onepassword.*",
  "sorcery",
]
ignore_missing_imports = true

[tool.pyright]
defineConstant = { PYPY = false }

[tool.pytest.ini_options]
minversion = "9.0"
filterwarnings = [
  "error",
  "ignore:Deprecated call to `pkg_resources.declare_namespace:DeprecationWarning",
]
addopts = "--cov-branch --cov-report xml:coverage.xml --cov-report term-missing:skip-covered --cov=src/configator"
testpaths = [
  "tests",
]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["A", "B", "C", "E", "F", "FBT", "I", "N", "RUF", "S"]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101", "S106"]

[dependency-groups]
dev = [
  "mypy>=1.18",
  "pytest>=9",
  "pytest-asyncio>=1.3",
  "pytest-cov>=3",
  "rich>=14",
  "ruff>=0.14",
]

[tool.uv]
environments = [
  "sys_platform == 'darwin'",
  "sys_platform == 'linux'",
]
package = true
